tests/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
coverage>=7.0.0
mock>=4.0.0
aioresponses>=0.7.0
httpx>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
yookassa>=2.3.0
aiogram>=3.0.0
supabase>=2.0.0 